+++ lesson4/src/components/NoChat/NoChat.js
@@ -1,9 +1,9 @@
 import React from "react"
 import { ChatUi } from "../ChatUi/ChatUi"
 
-export const NoChat = () => (
+export const NoChat = ({chats}) => (
     <>
-    <ChatUi chats='chats'/>
+    <ChatUi chats={chats}/>
     <span>Please select a chat</span>
     </>
     )
