+++ lesson4/src/components/Router/Router.js
@@ -3,9 +3,36 @@ import { BrowserRouter, Link, Switch, Route, useParams, Redirect } from 'react-r
 import Home from '../Home/Home';
 import { NoChat } from '../NoChat/NoChat';
 import { Profile } from '../Profile/Profile';
+import { AUTHORS } from '../constants';
+import { Chat } from '../Chat/Chat';
 
+const initialChats = {
+    chat1: {
+        messages: [{text: '', author: AUTHORS.human, id: 'chat1-1'}],
+        name: 'Rodion Chat',
+        id: 'chat1',
+},
+    chat2: {
+        messages: [{text: '', author: AUTHORS.human, id: 'chat2-1'}],
+        name: 'Mark Chat',
+        id: 'chat2',
+},
+    chat3: {name: 'Victory Chat', id: 'chat3', messages: []},
+}
 
 export const Router = () => {
+    const [chats, setChats] = useState(initialChats);
+
+    const handleSendMessage = useCallback((newMessage, chatId) => {
+        setChats({
+            ...chats,
+            [chatId]: {
+                ...chats[chatId],
+                messages: [...chats[chatId].messages, newMessage]
+            },
+        });
+     
+       }, [chats]);
 
     return (
         <BrowserRouter>
@@ -16,6 +43,10 @@ export const Router = () => {
                 <li>
                     <Link to="/profile">PROFILE</Link>
                 </li>
+                <li>
+                    <Link to="/chat">CHAT</Link>
+                </li>
+
             </ul>
 
             <Switch>
@@ -23,14 +54,17 @@ export const Router = () => {
                     <Profile match={data.match}/>
                 )}>
                 </Route>
-                <Route path="/home/:chatId?">
+                <Route path="/home">
                     <Home/>
-                    </Route>
+                </Route>
+                <Route path="/chat/:chatId?">
+                    <Chat chats={chats} onAddMessage={handleSendMessage}/>
+                </Route>
                 <Route path="/" exact>
                     <h2>WELCOME</h2>
                 </Route>
                 <Route path="/nochat">
-                    <NoChat/>
+                    <NoChat chats={chats}/>
                 </Route>
                 <Route path="*">
                     <h2>404</h2>
